+++ api/src/routes/modules/producto.js
@@ -1,0 +1,219 @@
+const { Router } = require('express');
+const axios = require('axios');
+
+const { Producto, Usuario, Favorito } = require("../../db");
+
+const bodyParser = require("body-parser");
+
+const router = Router();
+
+router.use(bodyParser.urlencoded({ extended: false }));
+
+
+//--------------------BEBIDAS-------------------------
+
+const getDataBase = async () => {
+    return await Producto.findAll();
+  };
+  router.get("/bebidasApi", async (req, res, next) => {
+    try {
+      const bebidasInfo = await axios.get(
+        `https://bebidas-efc61-default-rtdb.firebaseio.com/results.json`
+      );
+      const allBebidas = await bebidasInfo.data.map((e) => {
+        return e;
+      });
+      const allBebidasDb = await allBebidas.map((e) => {
+        Producto.create(e);
+      });
+  
+      res.json(allBebidas);
+    } catch (error) {
+      next(error);
+    }
+  });
+  
+  router.get("/bebidas", async (req, res, next) => {
+    try {
+      const { nombre } = req.query;
+      const dataInfo = await getDataBase();
+      if (nombre) {
+        const dataName = await dataInfo.filter((e) =>
+          e.nombre.toLowerCase().includes(nombre.toLowerCase())
+        );
+        if (!dataName.length) {
+          let error = [];
+          return res.json(error);
+        }
+        res.json(dataName);
+      } else {
+        res.json(dataInfo);
+      }
+    } catch (error) {
+      next(error);
+    }
+  });
+  
+  //--------------------BEBIDA-------------------------
+  
+
+router.post("/bebida", async (req, res) => {
+    let = { nombre, imagen, marca, ml, graduacion, descripcion, precio, stock } =
+      req.body;
+  
+    let [bebidaCreada, created] = await Producto.findOrCreate({
+      where: {
+        nombre: nombre,
+        imagen: imagen,
+        marca: marca,
+        descripcion: descripcion,
+        ml: ml,
+        graduacion: graduacion,
+        precio: precio,
+        stock: stock,
+      },
+    });
+    res.json(bebidaCreada);
+  });
+  
+  router.get("/bebida/:id", async (req, res) => {
+    let { id } = req.params;
+  
+    try {
+      let bebida = await Producto.findByPk(id);
+      res.status(200).json(bebida);
+    } catch (err) {
+      res.status(404);
+    }
+  });
+  
+  router.put("/bebida", async (req, res) => {
+    let { nombre, imagen, marca, ml, graduacion, descripcion, precio, stock } =
+      req.body;
+    let { id } = req.body;
+  
+    try {
+      const bebidaPut = await Producto.findOne({ where: { id: id } });
+  
+      await bebidaPut.update({
+        id: id,
+        nombre: nombre,
+        imagen: imagen,
+        marca: marca,
+        descripcion: descripcion,
+        ml: ml,
+        graduacion: graduacion,
+        precio: precio,
+        stock: stock,
+      });
+      res.json(bebidaPut);
+    } catch (err) {
+      console.log("error del glorioso catch. Amen");
+    }
+  });
+  
+  router.delete("/bebida/:id", async (req, res) => {
+    const { id } = req.params;
+  
+    const del = await Producto.destroy({
+      where: {
+        id: id,
+      },
+    });
+    return res.status(200).send("AL LOBBY");
+  });
+  
+  //-------------------BEBIDA FAVORITO------------------//
+
+// "id_prod":"eb9d3249-9eb9-4cc7-b562-8ced0d91e026",
+// "id_user": "S0aGACK7d3NCvnYNNdu9GSzTgrw2"
+router.post("/producto", async (req, res) => {
+    let { id_prod, id_user } = req.body;
+  
+    try {
+      let usuarioFavorito = await Usuario.findByPk(id_user, {});
+  
+      let productoFavorito = await Producto.findByPk(id_prod, {});
+  
+      // console.log(usuarioFavorito)
+      // console.log(productoFavorito)
+  
+    
+      usuarioFavorito.addUsuario(productoFavorito);
+      res.json(usuarioFavorito);
+  
+    } catch (err) {
+      console.log(err.message);
+    }
+  });
+  
+  router.get("/producto/favoritos", async (req, res) => {
+    let user = await Usuario.findOne({
+      include: {
+        model: Producto,
+        attributes: ["id", "nombre"],
+      },
+    });
+    console.log(user.productos, "ACA ESTOYYYYYYYYYYYYYYYYYYYYYYYYYYYYYY");
+    res.json(user);
+  });
+  
+  router.delete("/producto/favoritos", async (req, res) => {
+    let { id_prod, id_user } = req.body;
+  
+    let favBorrado = await Favorito.destroy({
+      where: {
+        usuarioId: id_user,
+        productoId: id_prod,
+      },
+    });
+  
+    res.json(Favorito);
+  });
+  
+  //////AQUI YACEN LOS RESTOS DE AUTENTICACION----RIP-AUTENTICACION----GRACIAS JONA </3----//////
+  //#region
+  
+  //   router.post('/usuario/login',  async (req, res) => {
+  //       const user ={
+  //           id,nombre,email
+  //       }=req.body
+  
+  //       jwt.sign({user},'secretkey',(err,token)=>{
+  //           res.json({token})
+  //       })
+  
+  //   })
+  
+  //   router.post('/usuario/posts', verifyToken, async (req, res) => {
+  
+  //    jwt.verify(req.token, 'secretkey',(error,authData) =>{
+  //        if(error){
+  //            res.sendStatus(403)
+  //        }else{
+  //            res.json({
+  //                mensaje:"Post fue creado",
+  //                authData
+  //            })
+  //        }
+  
+  //    })
+  
+  //   })
+  
+  // //Authorization: Bearer <token>
+  //   function verifyToken(req, res, next){
+  //       const bearerHeader = req.headers['authorization']
+  
+  //       if(typeof bearerHeader !== 'undefined'){
+  //           const bearerToken = bearerHeader.split(" ")[1];
+  //           req.token =bearerToken;
+  //           next();
+  //       }else{
+  //           res.sendStatus(403)
+  //       }
+  //   }
+  
+  //#endregion  //////////////////////////////////////////////////////////////////////
+  
+  module.exports = router;
