+++ api/src/routes/modules/review.js
@@ -1,0 +1,8 @@
+const { Router } = require('express');
+const axios = require('axios');
+
+const { Producto } = require("../../db");
+
+const router = Router();
+
+module.exports = router;
